refactor(jobs): extract register-job-handlers out of memory/ into a jobs/ composition root
The cross-domain job-handler registry is general worker infra, not memory-feature
code — it wires handlers across conversations/home/media/persistence/runtime/plugins.
Move it to a neutral jobs/ composition root so memory/ can become a plugin restricted
to its own files + plugin-api. Behavior byte-identical; worker.ts relocation deferred.

Part of memory-as-a-plugin Track 2.
