+++ .github/CODEOWNERS
@@ -10,6 +10,9 @@
 # Memory system (Sidd)
 assistant/src/memory/ @siddseethepalli
 
+# Background-job worker composition root (Sidd)
+assistant/src/jobs/ @siddseethepalli
+
 # Tool definitions and infrastructure (Sidd)
 assistant/src/tools/ @siddseethepalli
