+++ assistant/src/memory/__tests__/jobs-worker-v2-graph-trigger-embed.test.ts
@@ -67,13 +67,13 @@ const tmpWorkspace = mkdtempSync(
 const previousWorkspaceEnv = process.env.VELLUM_WORKSPACE_DIR;
 process.env.VELLUM_WORKSPACE_DIR = tmpWorkspace;
 
+import { registerMemoryJobHandlers } from "../../jobs/register-job-handlers.js";
 import { getMemoryDb } from "../../persistence/db-connection.js";
 import { initializeDb } from "../../persistence/db-init.js";
 import { _resetQdrantBreaker } from "../../persistence/embeddings/qdrant-circuit-breaker.js";
 import { enqueueMemoryJob } from "../../persistence/jobs-store.js";
 import { runMemoryJobsOnce } from "../../persistence/jobs-worker.js";
 import { memoryJobs } from "../../persistence/schema/index.js";
-import { registerMemoryJobHandlers } from "../register-job-handlers.js";
 
 describe("graph_trigger_embed under memory v2", () => {
   beforeAll(async () => {
