+++ assistant/src/memory/worker.ts
@@ -13,10 +13,10 @@
 import { existsSync, unlinkSync, writeFileSync } from "node:fs";
 
 import { getConfig } from "../config/loader.js";
+import { registerMemoryJobHandlers } from "../jobs/register-job-handlers.js";
 import { startInProcessMemoryJobsWorker } from "../persistence/jobs-worker.js";
 import { getLogger } from "../util/logger.js";
 import { getMemoryWorkerPidPath } from "../util/platform.js";
-import { registerMemoryJobHandlers } from "./register-job-handlers.js";
 
 const log = getLogger("memory-worker-process");
